+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.182 ] - 2025/11/05
+
+  * Added voice assist intent: `SpotifyPlusFavoriteAddRemove`.  This replaces 14 other intents (see below).
+  * Added voice assist intent: `SpotifyPlusGetNowPlayingInfo`.  This replaces 4 other intents (see below).
+  * Removed 14 voice assist intents related to favorite processing: `SpotifyPlusFavoriteAlbumAdd`, `SpotifyPlusFavoriteAlbumRemove`, `SpotifyPlusFavoriteArtistAdd`, `SpotifyPlusFavoriteArtistRemove`, `SpotifyPlusFavoriteAudiobookAdd`, `SpotifyPlusFavoriteAudiobookRemove`, `SpotifyPlusFavoritePlaylistAdd`, `SpotifyPlusFavoritePlaylistRemove`, `SpotifyPlusFavoritePodcastAdd`, `SpotifyPlusFavoritePodcastRemove`, `SpotifyPlusFavoritePodcastEpisodeAdd`, `SpotifyPlusFavoritePodcastEpisodeRemove`, `SpotifyPlusFavoriteTrackAdd`, `SpotifyPlusFavoriteTrackRemove`; manual removal from `config/custom_sentences` folder is required.
+  * Removed 4 voice assist intents related to nowplaying information processing: `SpotifyPlusNowPlayingInfoArtistBio`, `SpotifyPlusNowPlayingInfoAudiobook`, `SpotifyPlusNowPlayingInfoPodcast`, `SpotifyPlusNowPlayingInfoTrack`; manual removal from `config/custom_sentences` folder is required.
+
 ###### [ 1.0.181 ] - 2025/10/27
 
   * Updated `spotifywebapipython` package requirement to `spotifywebapipython>=1.0.254`.